Extract TestData as a top-level class
